move: MainActivity app 모듈 -> navigator 모듈 이동
