+++ feature/navigator/src/main/java/com/susu/feature/navigator/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.oksusu.susu
+package com.susu.feature.navigator
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -10,7 +10,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.oksusu.susu.ui.theme.SusuTheme
+import com.susu.core.designsystem.theme.SusuTheme
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
