--- feature/navigator/src/main/java/com/susu/feature/navigator/Sample.kt
@@ -1,1 +1,0 @@
-package com.susu.feature.navigator
